locust-loadtesting/docker-image/locust-tasks/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
ConfigArgParse==1.2.3
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==20.9.0
geventhttpclient==1.4.4
greenlet==0.4.17
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
locust==1.3.1
MarkupSafe==1.1.1
msgpack==1.0.0
psutil==5.7.3
pyzmq==18.0.1
requests==2.21.0
six==1.12.0
urllib3==1.24.2
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.1.2
google-oauth==1.0.1
google-auth==1.33.0